Externalize enums into seperate file
