+++ src/ABBAurora.h
@@ -1,6 +1,7 @@
 #ifndef ABBAurora_h
 #define ABBAurora_h
 #include <Arduino.h>
+#include <ABBAuroraEnums.h>
 #include <Wire.h>
 
 //RS485 control
@@ -11,7 +12,6 @@ class ABBAurora
 {
 private:
     int MaxAttempt = 1;
-    byte Address = 0;
     static byte TXPinControl;
     static HardwareSerial *serial;
 
@@ -34,6 +34,7 @@ private:
 public:
     bool SendStatus = false;
     bool ReceiveStatus = false;
+    byte Address = 0;
     byte ReceiveData[8];
 
     static void setup(HardwareSerial &serial, byte RXGpioPin, byte TXGpioPin, byte TXControllPin);
@@ -193,74 +194,4 @@ public:
     bool ReadSystemSerialNumberCentral();
 };
 
-enum DSP_VALUE_TYPE
-{
-    GRID_VOLTAGE = 1,
-    GRID_CURRENT = 2,
-    GRID_POWER = 3,
-    FREQUENCY = 4,
-    V_BULK = 5,
-    I_LEAK_DC_DC = 6,
-    I_LEAK_INVERTER = 7,
-    POWER_IN_1 = 8,
-    POWER_IN_2 = 9,
-    TEMPERATURE_INVERTER = 21,
-    TEMPERATURE_BOOSTER = 22,
-    V_IN_1 = 23,
-    I_IN_1 = 25,
-    V_IN_2 = 26,
-    I_IN_2 = 27,
-    DC_DC_GRID_VOLTAGE = 28,
-    DC_DC_GRID_FREQUENCY = 29,
-    ISOLATION_RESISTANCE = 30,
-    DC_DC_V_BULK = 31,
-    AVERAGE_GRID_VOLTAGE = 32,
-    V_BULK_MID = 33,
-    POWER_PEAK = 34,
-    POWER_PEAK_TODAY = 35,
-    GRID_VOLTAGE_NEUTRAL = 36,
-    WIND_GENERATOR_FREQENCY = 37,
-    GRID_VOLTAGE_NEUTRAL_PHASE = 38,
-    GRID_CURRENT_PHASE_R = 39,
-    GRID_CURRENT_PHASE_S = 40,
-    GRID_CURRENT_PHASE_T = 41,
-    FREQUENCY_PHASE_R = 42,
-    FREQUENCY_PHASE_S = 43,
-    FREQUENCY_PHASE_T = 44,
-    V_BULK_POSITIVE = 45,
-    V_BULK_NEGATIVE = 46,
-    TEMPERATURE_SUPERVISOR = 47,
-    TEMPERATURE_ALIM = 48,
-    TEMPERATURE_HEAT_SINK = 49,
-    TEMPERATURE_1 = 50,
-    TEMPERATURE_2 = 51,
-    TEMPERATURE_3 = 52,
-    FAN_SPEED_1 = 53,
-    FAN_SPEED_2 = 54,
-    FAN_SPEED_3 = 55,
-    FAN_SPEED_4 = 56,
-    FAN_SPEED_5 = 57,
-    POWER_SATURATION_LIMIT = 58,
-    V_PANEL_MICRO = 60,
-    GRID_VOLTAGE_PHASE_R = 61,
-    GRID_VOLTAGE_PHASE_S = 62,
-    GRID_VOLTAGE_PHASE_T = 63
-
-};
-
-enum DSP_GLOBAL
-{
-    GLOBAL_MESSUREMENT = 1,
-    MODULE_MESSUREMENT = 0,
-};
-
-enum CUMULATED_ENERGY_TYPE
-{
-    CURRENT_DAY = 0,
-    CURRENT_WEEK =0,
-    CURRENT_MONTH = 3,
-    CURRENT_YEAR = 4,
-    TOTAL = 5,
-    SINCE_RESET = 6
-};
 #endif
